+++ OutwardModsCommunicator/OutwardModsCommunicator.cs
@@ -25,7 +25,7 @@ namespace OutwardModsCommunicator
 
         public const string NAME = "Outward Mods Communicator";
 
-        public const string VERSION = "1.0.0";
+        public const string VERSION = "1.0.1";
 
         public static string xmlFilePath { get; set; } = "";
 
@@ -74,14 +74,15 @@ namespace OutwardModsCommunicator
             static void Postfix(ResourcesPrefabManager __instance)
             {
                 #if DEBUG
-                OMC.Log($"OMC@ResourcesPrefabManager_Load:\"{xmlFilePath}\"");
+                OMC.Log($"OMC@ResourcesPrefabManager_Load:\"{OMC.xmlFilePath}\"");
                 #endif
 
-                if (string.IsNullOrEmpty(xmlFilePath))
-                    return;
+                if (!string.IsNullOrEmpty(OMC.xmlFilePath))
+                    ConfigOverrideManager.OverrideConfigsFromFile(OMC.xmlFilePath);
 
-                ConfigOverrideManager.OverrideConfigsFromFile(xmlFilePath);
-                ConfigOverrideManager.OverrideConfigsFromFile(Path.Combine(GetProjectLocation(), "PlayerModsOverrides.xml"));
+                string playerOverridesLocation = Path.Combine(GetProjectLocation(), "PlayerModsOverrides.xml");
+                if (!string.IsNullOrEmpty(playerOverridesLocation))
+                    ConfigOverrideManager.OverrideConfigsFromFile(playerOverridesLocation);
             }
         }
     }