+++ Public/CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Changelog
 
+### Release 1.0.1 Version
+
+#### Fixed
+
+-   Made sure to load `PlayerModsOverrides.xml` even if `xmlFilePath` variable
+    is not provided from mods
+-   Changed readme description
+
 ### Release 1.0.0 Version
 
 -   Initial release.